Update import paths in JavaScript files, update lib-franklin.
