+++ 404.html
@@ -11,7 +11,7 @@
   <meta property="og:title" content="Page not found">
   <script src="/_src-lp/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
-    import { sampleRUM } from '/_src-lp/scripts/lib-franklin.js';
+    import { sampleRUM } from '/_src-lp/scripts/aem.js';
 
     window.addEventListener('load', () => {
       if (document.referrer) {
