+++ head.html
@@ -1,6 +1,6 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
 <meta name="robots" content="noindex,nofollow"/>
-<script src="/_src-lp/scripts/lib-franklin.js" type="module"></script>
+<script src="/_src-lp/scripts/aem.js" type="module"></script>
 <script src="/_src-lp/scripts/scripts.js" type="module"></script>
 
 <link rel="stylesheet" href="/_src-lp/styles/vendor/bootstrap.min.css"/>
